pyproject.toml
[project]
name = "rust-vs-python-rag-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.8.3",
    "hnswlib>=0.8.0",
    "httpx>=0.28.1",
    "hypothesis>=6.151.9",
    "langchain-text-splitters>=1.1.1",
    "llama-cpp-python>=0.3.16",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "onnxruntime>=1.24.4",
    "psutil>=7.2.2",
    "pytest>=9.0.2",
    "respx>=0.22.0",
    "sentence-transformers>=5.3.0",
    "tokenizers>=0.22.2",
    "tomli>=2.4.0",
    "waitress>=3.0.2",
]
